refactor: moved HomeScreen components to their own folder
